+++ pvinspect/analysis/__init__.py
@@ -1,0 +1,5 @@
+import defects
+import factory_models
+from default_plugins import _register_default_plugins
+
+_register_default_plugins()